圖片導入成功 顯示中文 修正食材清單不顯示的bug bug: android 14 不相容
