+++ app/src/main/java/com/example/idiotchefassistant/resultBlock/ResultRepository.kt
@@ -18,7 +18,7 @@ class ResultRepository {
 
     fun uploadData(newResults: Map<String, String>) {
         nowDatas.value = newResults
-        Log.i("ingredients name: ", newResults.keys.toString())
+//        Log.i("ingredients name: ", newResults.keys.toString())
     }
 
     fun getDatas(): Map<String, String>? {
