+++ app/src/main/res/layout/fragment_image_detail.xml
@@ -11,6 +11,6 @@
     android:id="@+id/fullImageView"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:scaleType="fitCenter" />
+    android:scaleType="centerInside" />
 
 </FrameLayout>
